+++ gradesa/src/app/lessons/exercises/dragdrop/itemtypes.js
